include patch for fixed-type followed by push-type input
This patch is a little ugly and maybe not even necessary.

It allows people to create (bad) inputs like this

    <!-- fixed -->
    <input name="foo[1]" value="a">

    <!-- push -->
    <input name="foo[]" value="b">
    <input name="foo[]" value="c">

Before Patch (bug):

    ['b', 'c']

After Patch (proper result):

    [null, 'a', 'b', 'c']

-----

However, even though this works, it's simply bad form design.

I'm leaving this on a branch for now because people should not be writing forms like this.

The original version (pre-patch) will function 100% for good form design.